requirements.txt
grpcio==1.38.1
grpcio-tools==1.36.1
ipdb
elasticsearch==7.12.0
elasticsearch-dsl==7.3.0
SQLAlchemy==1.4.19
PyMySQL==1.0.2
transformers==4.18.0
faiss==1.5.3
numpy==1.19.5
sentence-transformers
sentencepiece==0.1.96
psutil==5.8.0
memory-profiler
Flask==2.0.1
requests==2.25.1
cryptography
Pillow==8.2.0
nltk==3.6.2
protobuf==3.17.3
python-dateutil==2.8.1
torch==1.9.0
urllib3==1.26.5
vncorenlp==1.0.3
pyvi==0.1.1
# Backend
Django==3.2.13
djangorestframework==3.12.4
pandas
django-cors-headers